test: isolate x tld updater script tests
