+++ crates/runner/mitm-addon/tests/test_x_billing.py
@@ -11,7 +11,6 @@ emitted bucket must have a dev-seed row, etc.
 from __future__ import annotations
 
 import functools
-import importlib.util
 import json
 import pathlib
 import re
@@ -30,7 +29,6 @@ from usage.providers.connectors.x_billing import (
     classify_bucket,
     refine_bucket_with_body,
 )
-from usage.providers.connectors.x_tlds import IANA_TLD_VERSION, IANA_TLDS
 
 
 class _FirewallPermission(NamedTuple):
@@ -350,24 +348,6 @@ def _sample_path_for_pattern(pattern: str) -> str:
             segments.append(f"{match.group('prefix')}sample{match.group('suffix')}")
 
     return "/" + "/".join(segments)
-
-
-class TestTldSnapshot:
-    """Static invariants for the checked-in IANA TLD snapshot."""
-
-    def test_snapshot_has_version_and_expected_stable_entries(self):
-        assert IANA_TLD_VERSION
-        assert {"ai", "com", "dev", "museum", "xn--q9jyb4c"} <= IANA_TLDS
-
-    def test_snapshot_is_canonical(self):
-        addon_root = pathlib.Path(__file__).resolve().parent.parent
-        script = addon_root / "scripts" / "update-x-tlds.py"
-        spec = importlib.util.spec_from_file_location("update_x_tlds", script)
-        assert spec is not None
-        assert spec.loader is not None
-        module = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module)
-        assert module.check_generated() == 0
 
 
 class TestFirewallConsistency:
